Remove check for course default grading scheme
This is part 1 of 2, only this commit is meant to be hot fixed. In production we do not yet have the code to display what courses are using a grading scheme as default, so we should not be considering a grading scheme as "used" (and thus un-deletable) if users cannot view the courses they are used on.

Also re-orders checks for performance and only considers a grading scheme as used for an assignment if it has been graded with (the same logic Canvas has always used)

flag=archived_grading_schemes

test plan:
- Set GS as course default but do not use for grading an assignment
- Delete the GS

Change-Id: I71571ef83eca1b203db2cf543101cf891745243e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/355433
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Reviewed-by: Sleyder Zuleta <sleyder.zuleta@instructure.com>
QA-Review: Cameron Ray <cameron.ray@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>