fix overflow issue on account user page
When visiting a user with a long list of courses in the
"Enrollments" > "Courses" scrollable div, the overall page
height would expand to match the height of the list within
the scrollable div.

Applying `position: relative` to the <ul> element isolates it
and prevents it from improperly influencing the page height.
This adjustment ensures the browser respects the visible size
of the element, resolving the issue of the page appearing
excessively tall.

flag=none
closes FOO-4602

Test plan:
- visit an account user’s detail page with a long list of
  course enrollments
  - example: https://ocps.instructure.com/users/3311329
- EXPECT the page to not be affected by the DOM elements
  inside the scrollable enrollments div
- EXPECT the temporary enrollments div to not affect the
  page height
- EXPECT the accounts div to not affect the page height
- EXPECT the groups div to not affect the page height

[skip-crystalball]

Change-Id: I3e7afeb1ccb4f0bbce21629ca24f508283c8c4e0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/355892
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>