Spiff up doFetchApi for non-JSON responses
Fixes FOO-4748
flag=none

The doFetchApi effect was missing a bit of functionality.

A response body was always assumed to be JSON and fed
to JSON.stringify, so it couldn't be used for endpoints
that returned payloads that were not JSON.

Headers are now processed using Headers objects instead
of plain old JavaScript Objects. This makes the code
easier to reason about.

Headers can no longer be "snuck in" via fetchOpts.headers
and instead have to come in via the documented headers
property.

In the event of a FormData payload on a POST request,
we now make sure to NOT allow a Content-Type to be
specified. This is required per the MDN documentation
on multipart/form-data payloads.

Unit tests for all of the above were added or modified
as necessary. There were a lot of places in tests where
fetches were mocked to return a JSON but did not have
the Content-Type in the Response object set correctly.
That matters now! (and Rails will always get it right)

Test plan:
* automated tests pass

Change-Id: I10967c74199d85b8c753d5469c91d1797117b215
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/356663
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>