Set workflow_state when mirroring DKAB to LRAB
We weren't saving the DeveloerKeyAccountBinding's workflow_state
when creating a corresponding Lti::RegistrationAccountBinding.
This change fixes that.

It was likely not noticed because the default workflow_state is
"off," so after creating it, the user would then change the
workflow state to "on," which would trigger the update callback
and correctly set the workflow state that time around.

test plan:
- Create a site admin LTI developer key
- In a root account, turn that LTI key "on"
- In a rails console, check that `Lti::RegistrationAccountBinding.last.workflow_state`
  is "on."

flag = none

fixes INTEROP-8826

Change-Id: I21e4f19fe3f02342c075b51f2f56443abbb03f2c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/359570
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>