add edit copy permission to enhanced rubric assgmt
this commit adds a confirmation modal to the enhanced rubric assignment
page if the user does not have permissions to edit the rubric. The modal
lets the user know that editing the rubric will create a copy of the
rubric because they lack the permissions to do so. this commit also
fixes a caching issue with useQuery that was causing rubrics for courses
not to be refetched int the rubric search tray.

closes EVAL-4766
closes EVAL-4765

flag=enhanced_rubrics
flag=enhanced_rubrics_assignments

setup:
- in order to test the permissions, you must have 2 different rubrics
  created.
- Rubric A should not have any associations or assessments and
  at the account level.
- Rubric B should have at least 2 associations and assessments
  and be at the account level.

test plan:
- navigate to an assignment and attach Rubric A. Verify that clicking
  the edit button does not show the confirmation modal.
- reload the page and verify that clicking on the edit button still
  does not show the confirmation modal.
- remove the rubric and attach Rubric B. Verify that clicking the edit
  button shows the confirmation modal.
- verify clicking on Cancel just closes the modal and clicking on
  Confirm opens the edit rubric full screen modal.
- reload the page and verify that clicking on the edit button still
  shows the confirmation modal.

Change-Id: If5fffab912bcbbc4a4206f3fe60fa5c547ae15ab
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/360648
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Derek Williams <derek.williams@instructure.com>
QA-Review: Derek Williams <derek.williams@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>