Ensure allow-scripts is always disabled for user_content iframes
Iframes without a sandbox attribute have no restrictions on what they can
do. An iframe with an empty sandbox tag disallows everything the iframe
can do. The change here is to set up a list of options that we are OK
with iframes having and a list of options we definitely never want on
iframes. We will only alter the list of sandbox options if one is given
and remove the options we do not want. If no sandbox attribute is on a
given iframe we put one in with the options we are OK with.

fixes RCX-2431
flag=rce_transform_iframe_sandbox_attributes

test plan:
- Create some RCE content and edit the HTML to put in some custom iframes
- Here is an example of three different iframes:
<iframe id="iframe1" src="https://example.com"></iframe>
<iframe id="iframe2" sandbox="allow-scripts allow-same-origin"
  src="https://example.com"></iframe>
<iframe id="iframe3" sandbox="allow-scripts"
  src="https://example.com"></iframe>
- Save the RCE content
- Verify that no data was changed in the database with regards to the
sandbox attribute and options
- View the page and inspect the iframes
- Verify iframe1 has a sandbox attribute with all the options we allow.
- Verify iframe2 only has allow-same-origin
- Verify iframe3 has all the options we allow and does not include
allow-scripts.
- Add the malicious iframe referenced in the ticket to the page and save.
- View the page and verify it is blocked.

Change-Id: I51626d27d8f1b77e04cb8df792be312a5ad59880
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/361777
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>