avoid infinite recursion reporting a database error
when a database error happens and we try to report it to sentry,
we access the database for shard_id in extras_hash. if this raises
for the same reason, we end up with infinite recursion. so break
the cycle by rescuing the database error there

flag=none
closes FOO-4897

Change-Id: I8892d6b177414e0da33af091bacd85263480db5c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/362033
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>