Reliably determine that an image is svg
the issue here is that we need to reliably determine if
the image being rendered is an svg. We can't do that by looking
at the filename's extension because all we have is the canvas
files api URL. So now we fetch the image file first and inspect
its content-type.

This shouldn't have an impact on performance, since now the file
is in the browser's cache.

This also fixes a bug where if you replace an image in an ImageBlock,
the loading spinner never disappeared.

closes RCX-2645
flag=block_editor
test plan:
  - create a blank page
  - add an image block
  - add an image
  > inspect the dom. expect the image to be an img tag
  - replace the image with an svg image
  > expect the spinner to disappear when the image loads
  > inspect the dom. expect the svg to be inlined
  - if the svg has fill="currentColor", and you change the
    background color of the image's parent
  > expect the color of the svg to switch between black and white
    depending on the shade of the background color

Change-Id: I3b99fa9279415ffa027c5adde66608ae2033d0ad
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/362130
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>