fix student grades printing bugs
With the top_navigation_placement feature flag on,
the whole Canvas layout is wrapped into the InstUI
DrawerLayout component. Because of this layout,
printing out the grades summary for a student
does not work properly.
In the print preview, grades are only rendered
on the first page, and the other pages are blank.
We can fix this issue with CSS so that grades
will be rendered properly on all pages.

fixes INTEROP-8972
flag=top_navigation_placement

Test plan:
- Make sure the top_navigation_placement feature flag is on
- Create and publish assignments in a course
- Go to a student's grades summary page
(/courses/:course_id/grades/:user_id)
- Press Print Grades
  - Before this fix, the second page is blank
and the grades summary is not rendered correctly.
  - After this fix (make sure you run `yarn build:css`
so the change can take effect),
the content on the second page is there,
and the print preview layout is correct.

Change-Id: I9bab281fe864b4ae49ee40cd399b5f1b5eb8cd21
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/363721
Reviewed-by: Dustin Cowles <dustin.cowles@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>