updating status of sub_assignment shows in gradebook no refresh
fixes EGG-341
flag=discussion_checkpoints

Test Plan:
0a) Simple test plan is in the ticket.
0b) Specs pass with fix.
1a) create a graded dicussion with checkpoints.
 - courses/:course_id/discussion_topics/new
1b) go to the course gradebook
 - courses/:course_id/gradebook
1c) go to a students cell and open submissions tray.
1d) edit the status of on checkpoint(ex: reply to topic)
-- do not set as 'Excuse' because this works*
1e) With my ps this will update the cell color.
1f) Without my ps is will not. Further if you update
the cell again it will switch to the previous status color.

Notes:
On the front end we use response.data.all_submissions,
so thats what needs to be checked.
When the reassignment occurs in controller we need to reload
the orignal_sub for that to be reflected.
If a submission is excused it runs more callbacks such as grading,
which is why is successfully updates.

Change-Id: Ifb830b9edb75992bf363e55d67800e4c7b5ac519
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/364031
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Wilmer Corrales <wilmer.corrales@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>