fix module selector for unenrolled viewers
The original code that introduced the module selector for unenrolled
viewers had a naive you are either a student or not a student view. This
switches to checking to see if you have course content edit rights that
dictate if you see/use the module selectors. It also ensures if you are
not enrolled in the course in any context, and the feature flag is on,
you will see the filtered student view of the modules.

fixes CNVS-64497
flag=none

test plan:
 - Have a published course set to Institution visibility with multiple
   modules.
 - Enable the modules_teacher_module_selection and
   modules_student_module_selection feature flags
 - Go to the course and select a specific module to filter the view for
   both Student View and Teacher View
 - Access the course modules as a teacher and confirm the correct view
 - Access the course modules as a student and confirm the correct view
 - Access the course as a user not in the course, and ensure you the
   students view

Change-Id: Ifc4c4240fe15e71ad84a8656e1963f7e636709ee
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/365284
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>