Add support for fancy midnight to PeerReviewOptions.jsx
closes EGG-273
flag=none

Test plan:
 - Test passes
 - Try to create a new Discussion Topic
 - Click on Graded
 - In Peer Reviews, select the radio button "Automatically assign"
 - Use the date picker in "Assign Reviews" and pick a date.
 - Notice that the time is set to 11:59 PM.

NOTE: If you type the date and click enter, the time picker
will not update. This is an INSTUI bug and is happening in
Assign To and other places. They are already aware of it.
For reference: INSTUI-4406

Change-Id: Ib17e72ab7ae0ba6b2f33e589371c72e74087a12d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/365356
Reviewed-by: Angela Gomba <angela.gomba@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>