Improve difference detection in LtiOverlays
This commmit changes how the FE computes the values to send in overlays.

Previously, the FE would keep the values from the top configuration as
placeholders in many of the input fields in the registration wizard. The
user would be able to "overlay" values on top of those placeholders.
This wound up confusing users, as they would see the placeholder values
from the configuration, and thought that they "couldn't edit" them.

This changes the behavior so that the FE will preopulate the input
fields with the values from the configuration. However, in order to
maintain the cleanlines of the overlay (i.e. we don't want all the
values from the configuration duplicated in the overlay), we will only
send the values that are different from the configuration to the backend
when the user submits the registration.

test plan:

1. Create a new LTI 1.3 registration via JSON (pasting the JSON into the
   registration wizard)
2. Notice that the values in the form fields are prepopulated (not as
   placeholders) with the values from the JSON.

closes INTEROP-9025
flag=lti_registrations_next

Change-Id: I6eca0b220ddb1242217bdf86acd792cf4aa83d25
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/365684
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>