Fix flashing in the Registration Modal
This commit fixes an issue where the registration modal would flash with
different states when closing the modal. This was due to the state being
reset in lock-step with when the modal was closed. Since the modal
closing happened as a transition, the state would be reset before the
transition was complete, causing the flashing.

This commit fixes the issue by only resetting the state when the modal
opens, and not when it closes. It also gets rid of the "Success" state,
which only flashed for a moment before the modal closed.

closes INTEROP-9033
flag=lti_registrations_page

Change-Id: I7b899eeb98915d2b31efbcc35c56480e403ad448
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/366073
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>