Add requirement_type to modules API
The API does not currently indicate whether a module requires "all"
or "one" required item to be completed (it is only exposed via the
modules erb template).

This attribute is computed based off columns on the module object,
which is already loaded in memory.

This change does not require notice in API release notes since it is
adding an attribute to an existing API.

closes CLX-321
flag = none

Test plan:
 - Create a module with "Complete all" selected, and at least 1
   requirement.
 - Create another module with "Complete one" selected, and at least
   1 requirement.
 - Visit /api/v1/courses/:course_id/modules/
 - Expect the first module to have "requirement_type: all", and the
   second to have "requirement_type: one"

Change-Id: If8321e74287eb86dd133358262589751c27e9f80
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/366677
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Kevin Qiu <kqiu@instructure.com>
QA-Review: Kevin Qiu <kqiu@instructure.com>
Product-Review: Kevin Qiu <kqiu@instructure.com>