add anonymous ids to graphql
This patchset adds the anonymous ids to the submission object in graphql
and the ability to search by them. This also adds hiding the user/userId
on a submission from a user with the manage_grades permission when an
assignment is actively anonymous. This new user/userId behaviour is
behind a feature flag as this is not backwards compatible behaviour.

closes EVAL-3871
flag=graphql_honor_anonymous_grading

test plan:
 - Have a course with at least one student
 - Create an assignment with anonymous grading
 - As a teacher, query for submissions for that assignment making sure
   to include the the anonymousId, userId, and some values from the
   user.
   query {
     assignment(id: "187") {
       _id
       name
       submissionsConnection(filter: {includeUnsubmitted: true}) {
         nodes {
           _id
           anonymousId
           userId
           user {
             name
           }
         }
       }
     }
   }
 - Make sure the anonymousId is returned and the userId and user values
   are not returned.
 - Grade all the students assigned to that assignment
 - Run the same query again and make sure the userId and user values are
   returned.

Change-Id: Ifd3531be2422803084809f3659564fe65d475aba
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/368235
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>